Fix a few todos

+++ src/main/java/org/icatproject/ids/thread/Archiver.java
@@ -5,8 +5,8 @@ import java.nio.file.Path;
 
 import org.icatproject.ids.plugin.DsInfo;
 import org.icatproject.ids.plugin.MainStorageInterface;
-import org.icatproject.ids.util.PropertyHandler;
 import org.icatproject.ids.webservice.FiniteStateMachine;
+import org.icatproject.ids.webservice.PropertyHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -31,18 +31,22 @@ public class Archiver implements Runnable {
 	@Override
 	public void run() {
 
-		// TODO do nothing if preparation in progress
-
 		try {
-			// remove any files from the dataset cache
-			Path datasetCachePath = datasetCache.resolve(dsInfo.getFacilityName())
-					.resolve(dsInfo.getInvName()).resolve(dsInfo.getVisitId())
-					.resolve(dsInfo.getDsName());
+			String preparedId;
+			// Is the dataset in a preparation?
+			if ((preparedId = fsm.preparing(dsInfo)) != null) {
+				logger.debug("Archive of " + dsInfo + " skipped as needed by prepare " + preparedId);
+			} else {
+				// remove any files from the dataset cache
+				Path datasetCachePath = datasetCache.resolve(dsInfo.getFacilityName())
+						.resolve(dsInfo.getInvName()).resolve(dsInfo.getVisitId())
+						.resolve(dsInfo.getDsName());
 
-			Files.deleteIfExists(datasetCachePath);
+				Files.deleteIfExists(datasetCachePath);
 
-			mainStorageInterface.delete(dsInfo);
-			logger.debug("Archive of " + dsInfo + " completed");
+				mainStorageInterface.delete(dsInfo);
+				logger.debug("Archive of " + dsInfo + " completed");
+			}
 		} catch (Exception e) {
 			logger.error("Archive of " + dsInfo + " failed due to " + e.getMessage());
 		} finally {
